+++ bootstrap-dualselect.js
@@ -1,5 +1,20 @@
 'use strict';
 
+/*
+  Usage
+
+  <ts-dualselect label1="All items" label2="Selected items" ng-model="selected" ng-options="selectable" value="name"></ts-dualselect>
+
+  where:
+   label1 - label for left  select box
+   label2 - label for right select box
+   ng-options - elements that can be selected, displayed in left select box
+   ng-model - 'selected' elements, displayed in right select box
+   value - (optional) - if selectable elements are objects, value is name of the property. Value of the property
+                        will be displayed in the select
+
+
+ */
 var scripts = document.getElementsByTagName("script");
 var currentScriptPath = scripts[scripts.length-1].src;
 
@@ -25,16 +40,19 @@ angular.module('tstechly.angular-bootstrap-dualselect', []).directive('tsDualsel
       scope.label1 = attrs.label1;
       scope.label2 = attrs.label2;
       scope.selectedItems = scope.$parent[attrs.ngModel];
+      var valueAttribute = attrs.value
 
       scope.$on("dualselect.event:data-ready", function() {
         scope.selectable = scope.$parent[attrs.ngOptions];
         /* Add only items that are not on selectedItems list */
         for(var idx in scope.selectable) {
           var item = scope.selectable[idx];
-          scope.selectable.push(item.name);
-          if(!contains(scope.selectedItems, item.name) && item!==undefined && item.name!==undefined) {
-            scope.allItems.push(item.name);
-            scope.fAllItems.push(item.name);
+          var itemValue = valueAttribute!==undefined?item[valueAttribute]:item;
+
+          scope.selectable.push(itemValue);
+          if(!contains(scope.selectedItems, itemValue) && item!==undefined && itemValue!==undefined) {
+            scope.allItems.push(itemValue);
+            scope.fAllItems.push(itemValue);
           }
         }